feat: Update dependencies and add sharp library
The code changes include updating the dependencies in the package.json files for both the backend and UI, as well as adding the "sharp" library to both packages. This update will ensure that the project is using the latest stable versions of the dependencies and will enable the use of the "sharp" library for image processing.

Note: This is just a suggested commit message. Please feel free to modify it as needed.